requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==8.0.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
